refactor: split technology screen and trade order handler parts (#3878)
Extract TechnologyScreen top-bar toggle and tab bodies into library parts;
split Market tab direction vs quantity order handlers and production
breakdown table cell chrome. Document layout in repo-and-packages.md.
