+++ app/lib/features/game/screens/technology/technology_screen.dart
@@ -18,6 +18,9 @@ import '../../widgets/shell/shell_player_guarded_body.dart';
 import '../../widgets/technology/tech_tree_widget.dart';
 import '../../widgets/technology/technology_panel.dart';
 
+part 'technology_screen_top_bar.dart';
+part 'technology_screen_body.dart';
+
 /// Full-screen Technology screen with two tabs: Research Slots and Tech Tree.
 ///
 /// Dark editorial-monocle chrome per `SPEC/ui/technology-panel.md` § Top bar:
@@ -68,8 +71,6 @@ class TechnologyScreen extends ConsumerStatefulWidget {
   @override
   ConsumerState<TechnologyScreen> createState() => _TechnologyScreenState();
 }
-
-enum _TechnologyTab { slots, tree }
 
 class _TechnologyScreenState extends ConsumerState<TechnologyScreen> {
   _TechnologyTab _tab = _TechnologyTab.slots;
@@ -124,148 +125,5 @@ class _TechnologyScreenState extends ConsumerState<TechnologyScreen> {
         }
       },
     );
-  }
-}
-
-/// Slots / Tree toggle for the trailing slot of the technology top bar.
-///
-/// Implements the mockup `.tab-row` rule with two non-Material chip buttons
-/// painted in the dark editorial-monocle palette. Selected chip uses
-/// `--accent` border + accent-tinted background; unselected uses `--border`
-/// + transparent background. No Material `Chip` / `ChoiceChip` /
-/// `ToggleButtons` per the catalog ban.
-class _TechnologyTabToggle extends StatelessWidget {
-  const _TechnologyTabToggle({required this.selected, required this.onSelect});
-
-  final _TechnologyTab selected;
-  final void Function(_TechnologyTab next) onSelect;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisSize: MainAxisSize.min,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: <Widget>[
-        _TechnologyTabChip(
-          key: TechnologyScreen.slotsToggleKey,
-          // ignore: avoid_hardcoded_strings_in_widgets
-          label: 'Slots',
-          selected: selected == _TechnologyTab.slots,
-          onTap: () => onSelect(_TechnologyTab.slots),
-        ),
-        const SizedBox(width: 6),
-        _TechnologyTabChip(
-          key: TechnologyScreen.treeToggleKey,
-          // ignore: avoid_hardcoded_strings_in_widgets
-          label: 'Tree',
-          selected: selected == _TechnologyTab.tree,
-          onTap: () => onSelect(_TechnologyTab.tree),
-        ),
-      ],
-    );
-  }
-}
-
-class _TechnologyTabChip extends StatelessWidget {
-  const _TechnologyTabChip({
-    super.key,
-    required this.label,
-    required this.selected,
-    required this.onTap,
-  });
-
-  final String label;
-  final bool selected;
-  final VoidCallback onTap;
-
-  static const double _verticalPadding = 4;
-  static const double _horizontalPadding = 10;
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-    final TextStyle base =
-        theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12);
-    final Color borderColor = selected
-        ? EditorialMonoclePalette.accent
-        : EditorialMonoclePalette.border;
-    final Color labelColor = selected
-        ? EditorialMonoclePalette.accentBright
-        : EditorialMonoclePalette.muted;
-    final Color backgroundColor = selected
-        ? EditorialMonoclePalette.accent.withValues(alpha: 0.18)
-        : Colors.transparent;
-    return Semantics(
-      button: true,
-      selected: selected,
-      label: label,
-      child: Material(
-        color: Colors.transparent,
-        child: InkWell(
-          onTap: onTap,
-          child: AnimatedContainer(
-            duration: const Duration(milliseconds: 120),
-            curve: Curves.easeOut,
-            padding: const EdgeInsets.symmetric(
-              horizontal: _horizontalPadding,
-              vertical: _verticalPadding,
-            ),
-            decoration: BoxDecoration(
-              color: backgroundColor,
-              border: Border.all(color: borderColor, width: 1),
-            ),
-            child: Text(
-              label,
-              style: base.copyWith(
-                color: labelColor,
-                fontWeight: selected ? FontWeight.w600 : FontWeight.w500,
-                letterSpacing: 0.04,
-              ),
-              maxLines: 1,
-              overflow: TextOverflow.ellipsis,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _SlotsBody extends StatelessWidget {
-  const _SlotsBody({
-    required this.game,
-    required this.player,
-    this.currentOrders = const Orders(),
-    this.onOrdersChanged,
-  });
-
-  final Game game;
-  final Player player;
-  final Orders currentOrders;
-  final void Function(Orders orders)? onOrdersChanged;
-
-  @override
-  Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      padding: const EdgeInsets.all(CtSpacing.l),
-      child: TechnologyPanel(
-        game: game,
-        player: player,
-        currentOrders: currentOrders,
-        onOrdersChanged: onOrdersChanged,
-      ),
-    );
-  }
-}
-
-class _TreeBody extends StatelessWidget {
-  const _TreeBody({required this.game, required this.player});
-
-  final Game game;
-  final Player player;
-
-  @override
-  Widget build(BuildContext context) {
-    return TechTreeWidget(game: game, player: player);
   }
 }
