+++ app/lib/features/game/screens/technology/technology_screen_body.dart
@@ -1,0 +1,42 @@
+// Slots and Tree tab bodies for [TechnologyScreen].
+
+part of 'technology_screen.dart';
+
+class _SlotsBody extends StatelessWidget {
+  const _SlotsBody({
+    required this.game,
+    required this.player,
+    this.currentOrders = const Orders(),
+    this.onOrdersChanged,
+  });
+
+  final Game game;
+  final Player player;
+  final Orders currentOrders;
+  final void Function(Orders orders)? onOrdersChanged;
+
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      padding: const EdgeInsets.all(CtSpacing.l),
+      child: TechnologyPanel(
+        game: game,
+        player: player,
+        currentOrders: currentOrders,
+        onOrdersChanged: onOrdersChanged,
+      ),
+    );
+  }
+}
+
+class _TreeBody extends StatelessWidget {
+  const _TreeBody({required this.game, required this.player});
+
+  final Game game;
+  final Player player;
+
+  @override
+  Widget build(BuildContext context) {
+    return TechTreeWidget(game: game, player: player);
+  }
+}
